Add room type enum, fix imports
